+++ waves/scons_extensions.py
@@ -19,6 +19,7 @@ from waves._settings import _scons_substfile_suffix
 from waves._settings import _stdout_extension
 from waves._settings import _cd_action_prefix
 from waves._settings import _matlab_environment_extension
+from waves._settings import _sierra_environment_extension
 
 
 # TODO: Remove the **kwargs check and warning for v1.0.0 release
@@ -568,6 +569,92 @@ def abaqus_solver(program="abaqus", post_action=None, emitter=None, **kwargs):
         action=action,
         emitter=abaqus_emitter)
     return abaqus_solver_builder
+
+
+def _sierra_emitter(target, source, env):
+    """Appends the sierra builder target list with the builder managed targets
+
+    Appends ``target[0]``.stdout and ``target[0]``.env to the ``target`` list. The Sierra Builder requires
+    at least one target.
+
+    The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
+    a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
+    subdirectory, e.g. ``parameter_set1/target.ext``. When in doubt, provide the expected STDOUT redirected file as a
+    target, e.g. ``target[0].stdout``.
+
+    :param list target: The target file list of strings
+    :param list source: The source file list of SCons.Node.FS.File objects
+    :param SCons.Script.SConscript.SConsEnvironment env: The builder's SCons construction environment object
+
+    :return: target, source
+    :rtype: tuple with two lists
+    """
+    suffixes = [_stdout_extension, _sierra_environment_extension]
+    return _first_target_emitter(target, source, env, suffixes=suffixes)
+
+
+def sierra(program="sierra", application="adagio", post_action=None):
+    """Sierra SCons builder
+
+    This builder requires that the root input file is the first source in the list. The builder returned by this
+    function accepts all SCons Builder arguments and adds the keyword argument(s):
+
+    * ``sierra_options``: The Sierra command line options provided as a string.
+    * ``application_options``: The application (e.g. adagio) command line options provided as a string.
+
+    The first target determines the working directory for the builder's action, as shown in the action code snippet
+    below. The action changes the working directory to the first target's parent directory prior to executing the
+    journal file.
+
+    The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
+    a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
+    subdirectory, e.g. ``parameter_set1/target.ext``. When in doubt, provide the expected STDOUT redirected file as a
+    target, e.g. ``target[0].stdout``.
+
+    .. warning::
+
+       This is an experimental builder for Sierra support. The only emitted file is the application's version report in
+       ``TARGET[0].env`` and the ``TARGET[0].stdout`` redirected STDOUT and STDERR file. All relevant application output
+       files, e.g. ``genesis_output.e`` must be specified in the target list.
+
+    .. code-block::
+       :caption: SConstruct
+
+       import waves
+       env = Environment()
+       env.Append(BUILDERS={
+           "Sierra": waves.scons_extensions.sierra(),
+       })
+       Sierra(target=["output.e"], source=["input.i"])
+
+    .. code-block::
+       :caption: Sierra builder action
+
+       cd ${TARGET.dir.abspath} && ${program} ${sierra_options} ${application} ${application_options} -i ${SOURCE.file} > ${TARGET.filebase}.stdout 2>&1
+
+    :param str program: An absolute path or basename string for the Sierra program
+    :param str application: The string name for the Sierra application
+    :param list post_action: List of shell command string(s) to append to the builder's action list. Implemented to
+        allow post target modification or introspection, e.g. inspect the Sierra log for error keywords and throw a
+        non-zero exit code even if Sierra does not. Builder keyword variables are available for substitution in the
+        ``post_action`` action using the ``${}`` syntax. Actions are executed in the first target's directory as ``cd
+        ${TARGET.dir.abspath} && ${post_action}``.
+
+    :return: Sierra builder
+    :rtype: SCons.Builder.Builder
+    """
+    if not post_action:
+        post_action = []
+    action = [f"{_cd_action_prefix} {program} {application} --version > " \
+                  f"${{TARGET.filebase}}{_sierra_environment_extension}",
+              f"{_cd_action_prefix} {program} ${{sierra_options}} {application} ${{application_options}} " \
+                  f"-i ${{SOURCE.file}} > ${{TARGET.filebase}}{_stdout_extension} 2>&1"]
+    action.extend(_construct_post_action_list(post_action))
+    sierra_builder = SCons.Builder.Builder(
+        action=action,
+        emitter=_sierra_emitter
+    )
+    return sierra_builder
 
 
 def copy_substitute(source_list, substitution_dictionary=None, env=SCons.Environment.Environment(),